backend/services/rental-api/requirements.txt
# FastAPI 및 웹 서버
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 데이터베이스
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Pydantic
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# 날짜 처리
python-dateutil==2.8.2

# 인증 및 보안
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
httpx==0.25.1

# 보안 강화
slowapi==0.1.9
itsdangerous==2.1.2
sqlalchemy-utils==0.41.1
python-magic==0.4.27
Pillow==10.1.0
pythonjsonlogger==2.0.7
redis==5.0.1

# 개발 도구
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.0
pip-audit==2.6.1
